tune jrl; add install python bindings

+++ src/utils/print.cpp
@@ -3,7 +3,7 @@
 namespace atri {
 namespace utils {
 void print_problem(problem_ptr_t prob) {
-    for (int i = 0; i < field::num; i++) {
+    for (size_t i = 0; i < field::num; i++) {
         fmt::print("field : {}, \ttotal dim {}\n",
                    magic_enum::enum_name(static_cast<field_t>(i)),
                    prob->dim_[i]);
